+++ packages/graphiql/src/GraphiQL.tsx
@@ -10,21 +10,18 @@ import type {
   FC,
   ComponentPropsWithoutRef,
 } from 'react';
-import { Fragment, useState, useEffect, Children } from 'react';
+import { useState, useEffect, Children } from 'react';
 import {
   Button,
   ButtonGroup,
   ChevronDownIcon,
   ChevronUpIcon,
-  CopyIcon,
   Dialog,
   ExecuteButton,
   GraphiQLProvider,
   HeaderEditor,
   KeyboardShortcutIcon,
-  MergeIcon,
   PlusIcon,
-  PrettifyIcon,
   QueryEditor,
   ReloadIcon,
   ResponseEditor,
@@ -33,30 +30,25 @@ import {
   Tab,
   Tabs,
   Theme,
-  ToolbarButton,
   Tooltip,
   UnStyledButton,
-  copyQuery,
   useDragResize,
   useEditorStore,
   useExecutionStore,
-  mergeQuery,
   usePluginStore,
-  prettifyEditors,
   useSchemaStore,
   useStorage,
   useTheme,
   VariableEditor,
   WriteableEditorProps,
-  isMacOs,
   cn,
-  KEY_MAP,
 } from '@graphiql/react';
 import { HistoryStore, HISTORY_PLUGIN } from '@graphiql/plugin-history';
 import {
   DocExplorerStore,
   DOC_EXPLORER_PLUGIN,
 } from '@graphiql/plugin-doc-explorer';
+import { GraphiQLLogo, GraphiQLToolbar, GraphiQLFooter, ShortKeys } from './ui';
 
 /**
  * API docs for this live here:
@@ -138,7 +130,11 @@ const GraphiQL_: FC<GraphiQLProps> = ({
     >
       <HistoryStore maxHistoryLength={maxHistoryLength}>
         <DocExplorerStore>
-          <GraphiQLInterface {...interfaceProps}>{children}</GraphiQLInterface>
+          <Tooltip.Provider>
+            <GraphiQLInterface {...interfaceProps}>
+              {children}
+            </GraphiQLInterface>
+          </Tooltip.Provider>
         </DocExplorerStore>
       </HistoryStore>
     </GraphiQLProvider>
@@ -427,509 +423,368 @@ export const GraphiQLInterface: FC<GraphiQLInterfaceProps> = ({
   };
 
   return (
-    <Tooltip.Provider>
-      <div className={cn('graphiql-container', className)}>
-        <div className="graphiql-sidebar">
-          {plugins.map((plugin, index) => {
-            const isVisible = plugin === visiblePlugin;
-            const label = `${isVisible ? 'Hide' : 'Show'} ${plugin.title}`;
-            return (
-              <Tooltip key={plugin.title} label={label}>
-                <UnStyledButton
-                  type="button"
-                  className={cn(isVisible && 'active')}
-                  onClick={handlePluginClick}
-                  data-index={index}
-                  aria-label={label}
-                >
-                  <plugin.icon aria-hidden="true" />
-                </UnStyledButton>
-              </Tooltip>
-            );
-          })}
-          <Tooltip label="Re-fetch GraphQL schema">
-            <UnStyledButton
-              type="button"
-              disabled={isSchemaFetching}
-              onClick={introspect}
-              aria-label="Re-fetch GraphQL schema"
-              style={{ marginTop: 'auto' }}
-            >
-              <ReloadIcon
-                className={cn(isSchemaFetching && 'graphiql-spin')}
-                aria-hidden="true"
-              />
-            </UnStyledButton>
-          </Tooltip>
-          <Tooltip label="Open short keys dialog">
-            <UnStyledButton
-              type="button"
-              data-value="short-keys"
-              onClick={handleShowDialog}
-              aria-label="Open short keys dialog"
-            >
-              <KeyboardShortcutIcon aria-hidden="true" />
-            </UnStyledButton>
-          </Tooltip>
-          <Tooltip label="Open settings dialog">
-            <UnStyledButton
-              type="button"
-              data-value="settings"
-              onClick={handleShowDialog}
-              aria-label="Open settings dialog"
-            >
-              <SettingsIcon aria-hidden="true" />
-            </UnStyledButton>
-          </Tooltip>
-        </div>
-        <div className="graphiql-main">
-          <div
-            ref={pluginResize.firstRef}
-            className="graphiql-plugin"
-            style={{
-              // Make sure the container shrinks when containing long
-              // non-breaking texts
-              minWidth: '200px',
-            }}
-          >
-            {PluginContent ? <PluginContent /> : null}
-          </div>
-          {visiblePlugin && (
-            <div
-              className="graphiql-horizontal-drag-bar"
-              ref={pluginResize.dragBarRef}
-            />
-          )}
-          <div ref={pluginResize.secondRef} className="graphiql-sessions">
-            <div className="graphiql-session-header">
-              <Tabs
-                values={tabs}
-                onReorder={moveTab}
-                aria-label="Select active operation"
-                className="no-scrollbar"
+    <div className={cn('graphiql-container', className)}>
+      <div className="graphiql-sidebar">
+        {plugins.map((plugin, index) => {
+          const isVisible = plugin === visiblePlugin;
+          const label = `${isVisible ? 'Hide' : 'Show'} ${plugin.title}`;
+          return (
+            <Tooltip key={plugin.title} label={label}>
+              <UnStyledButton
+                type="button"
+                className={cn(isVisible && 'active')}
+                onClick={handlePluginClick}
+                data-index={index}
+                aria-label={label}
               >
-                {tabs.map((tab, index, arr) => (
-                  <Tab
-                    key={tab.id}
-                    value={tab}
-                    isActive={index === activeTabIndex}
+                <plugin.icon aria-hidden="true" />
+              </UnStyledButton>
+            </Tooltip>
+          );
+        })}
+        <Tooltip label="Re-fetch GraphQL schema">
+          <UnStyledButton
+            type="button"
+            disabled={isSchemaFetching}
+            onClick={introspect}
+            aria-label="Re-fetch GraphQL schema"
+            style={{ marginTop: 'auto' }}
+          >
+            <ReloadIcon
+              className={cn(isSchemaFetching && 'graphiql-spin')}
+              aria-hidden="true"
+            />
+          </UnStyledButton>
+        </Tooltip>
+        <Tooltip label="Open short keys dialog">
+          <UnStyledButton
+            type="button"
+            data-value="short-keys"
+            onClick={handleShowDialog}
+            aria-label="Open short keys dialog"
+          >
+            <KeyboardShortcutIcon aria-hidden="true" />
+          </UnStyledButton>
+        </Tooltip>
+        <Tooltip label="Open settings dialog">
+          <UnStyledButton
+            type="button"
+            data-value="settings"
+            onClick={handleShowDialog}
+            aria-label="Open settings dialog"
+          >
+            <SettingsIcon aria-hidden="true" />
+          </UnStyledButton>
+        </Tooltip>
+      </div>
+      <div className="graphiql-main">
+        <div
+          ref={pluginResize.firstRef}
+          className="graphiql-plugin"
+          style={{
+            // Make sure the container shrinks when containing long
+            // non-breaking texts
+            minWidth: '200px',
+          }}
+        >
+          {PluginContent ? <PluginContent /> : null}
+        </div>
+        {visiblePlugin && (
+          <div
+            className="graphiql-horizontal-drag-bar"
+            ref={pluginResize.dragBarRef}
+          />
+        )}
+        <div ref={pluginResize.secondRef} className="graphiql-sessions">
+          <div className="graphiql-session-header">
+            <Tabs
+              values={tabs}
+              onReorder={moveTab}
+              aria-label="Select active operation"
+              className="no-scrollbar"
+            >
+              {tabs.map((tab, index, arr) => (
+                <Tab
+                  key={tab.id}
+                  value={tab}
+                  isActive={index === activeTabIndex}
+                >
+                  <Tab.Button
+                    aria-controls="graphiql-session"
+                    id={`graphiql-session-tab-${index}`}
+                    title={tab.title}
+                    onClick={handleTabClick}
                   >
-                    <Tab.Button
-                      aria-controls="graphiql-session"
-                      id={`graphiql-session-tab-${index}`}
-                      title={tab.title}
-                      onClick={handleTabClick}
-                    >
-                      {tab.title}
-                    </Tab.Button>
-                    {arr.length > 1 && <Tab.Close onClick={handleTabClose} />}
-                  </Tab>
-                ))}
-              </Tabs>
-              <Tooltip label="New tab">
+                    {tab.title}
+                  </Tab.Button>
+                  {arr.length > 1 && <Tab.Close onClick={handleTabClose} />}
+                </Tab>
+              ))}
+            </Tabs>
+            <Tooltip label="New tab">
+              <UnStyledButton
+                type="button"
+                className="graphiql-tab-add"
+                onClick={addTab}
+                aria-label="New tab"
+              >
+                <PlusIcon aria-hidden="true" />
+              </UnStyledButton>
+            </Tooltip>
+            {logo}
+          </div>
+          <div
+            role="tabpanel"
+            id="graphiql-session" // used by aria-controls="graphiql-session"
+            aria-labelledby={`${TAB_CLASS_PREFIX}${activeTabIndex}`}
+          >
+            <div className="graphiql-editors" ref={editorResize.firstRef}>
+              <section
+                className="graphiql-query-editor"
+                aria-label="Query Editor"
+                ref={editorToolsResize.firstRef}
+              >
+                <QueryEditor
+                  editorTheme={editorTheme}
+                  keyMap={keyMap}
+                  onClickReference={onClickReference}
+                  onEdit={onEditQuery}
+                  readOnly={readOnly}
+                />
+                <div
+                  className="graphiql-toolbar"
+                  role="toolbar"
+                  aria-label="Editor Commands"
+                >
+                  <ExecuteButton />
+                  {toolbar}
+                </div>
+              </section>
+
+              <div
+                ref={editorToolsResize.dragBarRef}
+                className="graphiql-editor-tools"
+              >
                 <UnStyledButton
                   type="button"
-                  className="graphiql-tab-add"
-                  onClick={addTab}
-                  aria-label="New tab"
+                  className={cn(
+                    activeSecondaryEditor === 'variables' &&
+                      editorToolsResize.hiddenElement !== 'second' &&
+                      'active',
+                  )}
+                  onClick={handleToolsTabClick}
+                  data-name="variables"
                 >
-                  <PlusIcon aria-hidden="true" />
+                  Variables
                 </UnStyledButton>
-              </Tooltip>
-              {logo}
-            </div>
-            <div
-              role="tabpanel"
-              id="graphiql-session" // used by aria-controls="graphiql-session"
-              aria-labelledby={`${TAB_CLASS_PREFIX}${activeTabIndex}`}
-            >
-              <div className="graphiql-editors" ref={editorResize.firstRef}>
-                <section
-                  className="graphiql-query-editor"
-                  aria-label="Query Editor"
-                  ref={editorToolsResize.firstRef}
-                >
-                  <QueryEditor
-                    editorTheme={editorTheme}
-                    keyMap={keyMap}
-                    onClickReference={onClickReference}
-                    onEdit={onEditQuery}
-                    readOnly={readOnly}
-                  />
-                  <div
-                    className="graphiql-toolbar"
-                    role="toolbar"
-                    aria-label="Editor Commands"
-                  >
-                    <ExecuteButton />
-                    {toolbar}
-                  </div>
-                </section>
-
-                <div
-                  ref={editorToolsResize.dragBarRef}
-                  className="graphiql-editor-tools"
-                >
+                {isHeadersEditorEnabled && (
                   <UnStyledButton
                     type="button"
                     className={cn(
-                      activeSecondaryEditor === 'variables' &&
+                      activeSecondaryEditor === 'headers' &&
                         editorToolsResize.hiddenElement !== 'second' &&
                         'active',
                     )}
                     onClick={handleToolsTabClick}
-                    data-name="variables"
+                    data-name="headers"
                   >
-                    Variables
+                    Headers
                   </UnStyledButton>
-                  {isHeadersEditorEnabled && (
-                    <UnStyledButton
-                      type="button"
-                      className={cn(
-                        activeSecondaryEditor === 'headers' &&
-                          editorToolsResize.hiddenElement !== 'second' &&
-                          'active',
-                      )}
-                      onClick={handleToolsTabClick}
-                      data-name="headers"
-                    >
-                      Headers
-                    </UnStyledButton>
-                  )}
+                )}
 
-                  <Tooltip
-                    label={
+                <Tooltip
+                  label={
+                    editorToolsResize.hiddenElement === 'second'
+                      ? 'Show editor tools'
+                      : 'Hide editor tools'
+                  }
+                >
+                  <UnStyledButton
+                    type="button"
+                    onClick={toggleEditorTools}
+                    aria-label={
                       editorToolsResize.hiddenElement === 'second'
                         ? 'Show editor tools'
                         : 'Hide editor tools'
                     }
+                    className="graphiql-toggle-editor-tools"
                   >
-                    <UnStyledButton
-                      type="button"
-                      onClick={toggleEditorTools}
-                      aria-label={
-                        editorToolsResize.hiddenElement === 'second'
-                          ? 'Show editor tools'
-                          : 'Hide editor tools'
-                      }
-                      className="graphiql-toggle-editor-tools"
-                    >
-                      {editorToolsResize.hiddenElement === 'second' ? (
-                        <ChevronUpIcon
-                          className="graphiql-chevron-icon"
-                          aria-hidden="true"
-                        />
-                      ) : (
-                        <ChevronDownIcon
-                          className="graphiql-chevron-icon"
-                          aria-hidden="true"
-                        />
-                      )}
-                    </UnStyledButton>
-                  </Tooltip>
-                </div>
+                    {editorToolsResize.hiddenElement === 'second' ? (
+                      <ChevronUpIcon
+                        className="graphiql-chevron-icon"
+                        aria-hidden="true"
+                      />
+                    ) : (
+                      <ChevronDownIcon
+                        className="graphiql-chevron-icon"
+                        aria-hidden="true"
+                      />
+                    )}
+                  </UnStyledButton>
+                </Tooltip>
+              </div>
 
-                <section
-                  className="graphiql-editor-tool"
-                  aria-label={
-                    activeSecondaryEditor === 'variables'
-                      ? 'Variables'
-                      : 'Headers'
-                  }
-                  ref={editorToolsResize.secondRef}
-                >
-                  <VariableEditor
+              <section
+                className="graphiql-editor-tool"
+                aria-label={
+                  activeSecondaryEditor === 'variables'
+                    ? 'Variables'
+                    : 'Headers'
+                }
+                ref={editorToolsResize.secondRef}
+              >
+                <VariableEditor
+                  editorTheme={editorTheme}
+                  isHidden={activeSecondaryEditor !== 'variables'}
+                  keyMap={keyMap}
+                  onEdit={onEditVariables}
+                  onClickReference={onClickReference}
+                  readOnly={readOnly}
+                />
+                {isHeadersEditorEnabled && (
+                  <HeaderEditor
                     editorTheme={editorTheme}
-                    isHidden={activeSecondaryEditor !== 'variables'}
+                    isHidden={activeSecondaryEditor !== 'headers'}
                     keyMap={keyMap}
-                    onEdit={onEditVariables}
-                    onClickReference={onClickReference}
+                    onEdit={onEditHeaders}
                     readOnly={readOnly}
                   />
-                  {isHeadersEditorEnabled && (
-                    <HeaderEditor
-                      editorTheme={editorTheme}
-                      isHidden={activeSecondaryEditor !== 'headers'}
-                      keyMap={keyMap}
-                      onEdit={onEditHeaders}
-                      readOnly={readOnly}
-                    />
-                  )}
-                </section>
-              </div>
+                )}
+              </section>
+            </div>
 
-              <div
-                className="graphiql-horizontal-drag-bar"
-                ref={editorResize.dragBarRef}
+            <div
+              className="graphiql-horizontal-drag-bar"
+              ref={editorResize.dragBarRef}
+            />
+
+            <div className="graphiql-response" ref={editorResize.secondRef}>
+              {isExecutionFetching && <Spinner />}
+              <ResponseEditor
+                editorTheme={editorTheme}
+                responseTooltip={responseTooltip}
+                keyMap={keyMap}
               />
-
-              <div className="graphiql-response" ref={editorResize.secondRef}>
-                {isExecutionFetching && <Spinner />}
-                <ResponseEditor
-                  editorTheme={editorTheme}
-                  responseTooltip={responseTooltip}
-                  keyMap={keyMap}
-                />
-                {footer}
-              </div>
+              {footer}
             </div>
           </div>
         </div>
-        <Dialog
-          open={showDialog === 'short-keys'}
-          onOpenChange={handleOpenShortKeysDialog}
-        >
-          <div className="graphiql-dialog-header">
-            <Dialog.Title className="graphiql-dialog-title">
-              Short Keys
-            </Dialog.Title>
-            <Dialog.Close />
-          </div>
-          <div className="graphiql-dialog-section">
-            <ShortKeys keyMap={keyMap} />
-          </div>
-        </Dialog>
-        <Dialog
-          open={showDialog === 'settings'}
-          onOpenChange={handleOpenSettingsDialog}
-        >
-          <div className="graphiql-dialog-header">
-            <Dialog.Title className="graphiql-dialog-title">
-              Settings
-            </Dialog.Title>
-            <Dialog.Close />
-          </div>
-          {showPersistHeadersSettings ? (
-            <div className="graphiql-dialog-section">
-              <div>
-                <div className="graphiql-dialog-section-title">
-                  Persist headers
-                </div>
-                <div className="graphiql-dialog-section-caption">
-                  Save headers upon reloading.{' '}
-                  <span className="graphiql-warning-text">
-                    Only enable if you trust this device.
-                  </span>
-                </div>
-              </div>
-              <ButtonGroup>
-                <Button
-                  type="button"
-                  id="enable-persist-headers"
-                  className={cn(shouldPersistHeaders && 'active')}
-                  data-value="true"
-                  onClick={handlePersistHeaders}
-                >
-                  On
-                </Button>
-                <Button
-                  type="button"
-                  id="disable-persist-headers"
-                  className={cn(!shouldPersistHeaders && 'active')}
-                  onClick={handlePersistHeaders}
-                >
-                  Off
-                </Button>
-              </ButtonGroup>
-            </div>
-          ) : null}
-          {!forcedTheme && (
-            <div className="graphiql-dialog-section">
-              <div>
-                <div className="graphiql-dialog-section-title">Theme</div>
-                <div className="graphiql-dialog-section-caption">
-                  Adjust how the interface appears.
-                </div>
-              </div>
-              <ButtonGroup>
-                <Button
-                  type="button"
-                  className={cn(theme === null && 'active')}
-                  onClick={handleChangeTheme}
-                >
-                  System
-                </Button>
-                <Button
-                  type="button"
-                  className={cn(theme === 'light' && 'active')}
-                  data-theme="light"
-                  onClick={handleChangeTheme}
-                >
-                  Light
-                </Button>
-                <Button
-                  type="button"
-                  className={cn(theme === 'dark' && 'active')}
-                  data-theme="dark"
-                  onClick={handleChangeTheme}
-                >
-                  Dark
-                </Button>
-              </ButtonGroup>
-            </div>
-          )}
+      </div>
+      <Dialog
+        open={showDialog === 'short-keys'}
+        onOpenChange={handleOpenShortKeysDialog}
+      >
+        <div className="graphiql-dialog-header">
+          <Dialog.Title className="graphiql-dialog-title">
+            Short Keys
+          </Dialog.Title>
+          <Dialog.Close />
+        </div>
+        <div className="graphiql-dialog-section">
+          <ShortKeys keyMap={keyMap} />
+        </div>
+      </Dialog>
+      <Dialog
+        open={showDialog === 'settings'}
+        onOpenChange={handleOpenSettingsDialog}
+      >
+        <div className="graphiql-dialog-header">
+          <Dialog.Title className="graphiql-dialog-title">
+            Settings
+          </Dialog.Title>
+          <Dialog.Close />
+        </div>
+        {showPersistHeadersSettings ? (
           <div className="graphiql-dialog-section">
             <div>
-              <div className="graphiql-dialog-section-title">Clear storage</div>
+              <div className="graphiql-dialog-section-title">
+                Persist headers
+              </div>
               <div className="graphiql-dialog-section-caption">
-                Remove all locally stored data and start fresh.
+                Save headers upon reloading.{' '}
+                <span className="graphiql-warning-text">
+                  Only enable if you trust this device.
+                </span>
               </div>
             </div>
-            <Button
-              type="button"
-              state={clearStorageStatus || undefined}
-              disabled={clearStorageStatus === 'success'}
-              onClick={handleClearData}
-            >
-              {{
-                success: 'Cleared data',
-                error: 'Failed',
-              }[clearStorageStatus!] || 'Clear data'}
-            </Button>
+            <ButtonGroup>
+              <Button
+                type="button"
+                id="enable-persist-headers"
+                className={cn(shouldPersistHeaders && 'active')}
+                data-value="true"
+                onClick={handlePersistHeaders}
+              >
+                On
+              </Button>
+              <Button
+                type="button"
+                id="disable-persist-headers"
+                className={cn(!shouldPersistHeaders && 'active')}
+                onClick={handlePersistHeaders}
+              >
+                Off
+              </Button>
+            </ButtonGroup>
           </div>
-        </Dialog>
-      </div>
-    </Tooltip.Provider>
-  );
-};
-
-function withMacOS(key: string) {
-  return isMacOs ? key.replace('Ctrl', '⌘') : key;
-}
-
-const SHORT_KEYS = Object.entries({
-  'Search in editor': withMacOS(KEY_MAP.searchInEditor[0]),
-  'Search in documentation': withMacOS(KEY_MAP.searchInDocs[0]),
-  'Execute query': withMacOS(KEY_MAP.runQuery[0]),
-  'Prettify editors': KEY_MAP.prettify[0],
-  'Merge fragments definitions into operation definition':
-    KEY_MAP.mergeFragments[0],
-  'Copy query': KEY_MAP.copyQuery[0],
-  'Re-fetch schema using introspection': KEY_MAP.refetchSchema[0],
-});
-
-interface ShortKeysProps {
-  /** @default 'sublime' */
-  keyMap?: string;
-}
-
-const ShortKeys: FC<ShortKeysProps> = ({ keyMap = 'sublime' }) => {
-  return (
-    <div>
-      <table className="graphiql-table">
-        <thead>
-          <tr>
-            <th>Short Key</th>
-            <th>Function</th>
-          </tr>
-        </thead>
-        <tbody>
-          {SHORT_KEYS.map(([title, keys]) => (
-            <tr key={title}>
-              <td>
-                {keys.split('-').map((key, index, array) => (
-                  <Fragment key={key}>
-                    <code className="graphiql-key">{key}</code>
-                    {index !== array.length - 1 && ' + '}
-                  </Fragment>
-                ))}
-              </td>
-              <td>{title}</td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-      <p>
-        The editors use{' '}
-        <a
-          href="https://codemirror.net/5/doc/manual.html#keymaps"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          CodeMirror Key Maps
-        </a>{' '}
-        that add more short keys. This instance of Graph<em>i</em>QL uses{' '}
-        <code>{keyMap}</code>.
-      </p>
+        ) : null}
+        {!forcedTheme && (
+          <div className="graphiql-dialog-section">
+            <div>
+              <div className="graphiql-dialog-section-title">Theme</div>
+              <div className="graphiql-dialog-section-caption">
+                Adjust how the interface appears.
+              </div>
+            </div>
+            <ButtonGroup>
+              <Button
+                type="button"
+                className={cn(theme === null && 'active')}
+                onClick={handleChangeTheme}
+              >
+                System
+              </Button>
+              <Button
+                type="button"
+                className={cn(theme === 'light' && 'active')}
+                data-theme="light"
+                onClick={handleChangeTheme}
+              >
+                Light
+              </Button>
+              <Button
+                type="button"
+                className={cn(theme === 'dark' && 'active')}
+                data-theme="dark"
+                onClick={handleChangeTheme}
+              >
+                Dark
+              </Button>
+            </ButtonGroup>
+          </div>
+        )}
+        <div className="graphiql-dialog-section">
+          <div>
+            <div className="graphiql-dialog-section-title">Clear storage</div>
+            <div className="graphiql-dialog-section-caption">
+              Remove all locally stored data and start fresh.
+            </div>
+          </div>
+          <Button
+            type="button"
+            state={clearStorageStatus || undefined}
+            disabled={clearStorageStatus === 'success'}
+            onClick={handleClearData}
+          >
+            {{
+              success: 'Cleared data',
+              error: 'Failed',
+            }[clearStorageStatus!] || 'Clear data'}
+          </Button>
+        </div>
+      </Dialog>
     </div>
   );
-};
-
-const defaultGraphiqlLogo = (
-  <a
-    className="graphiql-logo-link"
-    href="https://github.com/graphql/graphiql"
-    target="_blank"
-    rel="noreferrer"
-  >
-    Graph
-    <em>i</em>
-    QL
-  </a>
-);
-
-// Configure the UI by providing this Component as a child of GraphiQL.
-const GraphiQLLogo: FC<{ children?: ReactNode }> = ({
-  children = defaultGraphiqlLogo,
-}) => {
-  return <div className="graphiql-logo">{children}</div>;
-};
-
-const DefaultToolbarRenderProps: FC<{
-  prettify: ReactNode;
-  copy: ReactNode;
-  merge: ReactNode;
-}> = ({ prettify, copy, merge }) => (
-  <>
-    {prettify}
-    {merge}
-    {copy}
-  </>
-);
-
-// Configure the UI by providing this Component as a child of GraphiQL.
-const GraphiQLToolbar: FC<{
-  children?: typeof DefaultToolbarRenderProps;
-}> = ({ children = DefaultToolbarRenderProps }) => {
-  if (typeof children !== 'function') {
-    throw new TypeError(
-      'The `GraphiQL.Toolbar` component requires a render prop function as its child.',
-    );
-  }
-
-  const prettify = (
-    <ToolbarButton
-      onClick={prettifyEditors}
-      label={`Prettify query (${KEY_MAP.prettify[0]})`}
-    >
-      <PrettifyIcon className="graphiql-toolbar-icon" aria-hidden="true" />
-    </ToolbarButton>
-  );
-
-  const merge = (
-    <ToolbarButton
-      onClick={mergeQuery}
-      label={`Merge fragments into query (${KEY_MAP.mergeFragments[0]})`}
-    >
-      <MergeIcon className="graphiql-toolbar-icon" aria-hidden="true" />
-    </ToolbarButton>
-  );
-
-  const copy = (
-    <ToolbarButton
-      onClick={copyQuery}
-      label={`Copy query (${KEY_MAP.copyQuery[0]})`}
-    >
-      <CopyIcon className="graphiql-toolbar-icon" aria-hidden="true" />
-    </ToolbarButton>
-  );
-
-  return children({ prettify, copy, merge });
-};
-
-// Configure the UI by providing this Component as a child of GraphiQL.
-const GraphiQLFooter: FC<{ children: ReactNode }> = ({ children }) => {
-  return <div className="graphiql-footer">{children}</div>;
 };
 
 function getChildComponentType(child: ReactNode) {
