+++ packages/graphiql/src/ui/index.tsx
@@ -1,0 +1,4 @@
+export { GraphiQLFooter } from './footer';
+export { GraphiQLLogo } from './logo';
+export { ShortKeys } from './short-keys';
+export { GraphiQLToolbar } from './toolbar';
